requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
black>=23.0.0  # for code formatting
flake8>=6.0.0  # for linting
google-generativeai>=0.3.0  # for Gemini API
python-dotenv>=1.0.0  # for environment variables
aiohttp>=3.8.0  # for async HTTP requests
tqdm>=4.65.0  # for progress bars
pytest-asyncio>=0.21.0 